requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Reuploaded==1.4.0
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
Pillow==10.1.0
Werkzeug==3.0.1
WTForms==3.1.1
